sumopinrisk/requirements.txt
aiohttp==3.1.3
asn1crypto==0.24.0
astroid==1.6.3
async-timeout==2.0.1
attrs==17.4.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.2.2
Django==2.0.2
django-excel-response==2.0.3
djangorestframework==3.8.2
et-xmlfile==1.0.1
idna==2.6
idna-ssl==1.0.1
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.2.0
openpyxl==2.4.11
pycparser==2.18
pylint==1.8.4
pylint-django==0.11
pylint-plugin-utils==0.2.6
pyOpenSSL==17.5.0
PySocks==1.6.8
pytz==2018.4
requestes==0.0.1
requests==2.18.4
six==1.11.0
sqlanydb==1.0.9
urllib3==1.22
win-inet-pton==1.0.1
wincertstore==0.2
wrapt==1.10.11
yarl==1.1.1
